Completed changes
